on a fait les fonctionnalite pour les classes c'est affichaes des classes avec ses informations par des card pour la dashboard  est aussi on a perpe les service cote front-end est aussi le foncitionnalite de filtrage  par niveaux  est aussi la fonctionnalite pour ajouter une nouvelle classes est auussli le fait de vois les details chaque classes son prof est ses eleves est ses details

+++ EcoSmart/app/Services/ServiceNiveau.php
@@ -8,19 +8,19 @@ use App\Models\Niveau;
 class ServiceNiveau {
     
     public function getAll(){
-
         return Niveau::all();
     }
     public function create(array $data): Niveau {
         return Niveau::create($data);
     }
-
     public function modifier(Niveau $niveau, array $data){
         $niveau->update($data);
         return $niveau;
     }
-
     public function supprimer(Niveau $niveau){
+        if($niveau->classes()->exists()){
+            throw new \Exception("cette niveaux a des classes");
+        }
         $niveau->delete();
         return true;
     }
